# net/packet: remove the custom IP4/IP6 types in favor of netaddr.IP.