@danderson 2020-12-20
Note to reviewers: I'm out starting tonight until next year. This is ready to merge when you're happy with the state of alloc-free netaddr. This PR points go.mod at the current tip of the bradfitz/rep24_zoneptr branch, which is necessary to pass tests (otherwise, we fail the allocation-freedom tests).