@bradfitz 2020-12-22
Rebased.